requirements.txt
ccxt
pandas
pandas-ta
pyyaml
streamlit
